+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 4.2.6 (unreleased)
 
 
-- Nothing changed yet.
+- Support `context` in `rephrase()`
 
 
 ## 4.2.5 (2024-10-17)